refactor(db-schema): file-organization sweep over the models
Limit constants move out of the entity and payload files into the
services constants they belong to: the message, mention and moderation
note lengths join `services/message/constants.ts`, the embed and
webhook bounds get `services/message/webhook/constants.ts`, and the
resource tag bounds join `services/resource/constants.ts`. The
sanitized message schema moves out of `BaseMessageEntity` into its own
model file since two scheduled-message payloads share it.
