+++ packages/db-schema/src/models/message/ModerationNoteEntity.ts
@@ -6,10 +6,9 @@ import { AzureEntity, createAzureEntitySchema } from "#src/models/azure/table/Az
 import { reverseTickedTimestampSchema } from "#src/models/azure/table/reverseTickedTimestampSchema";
 import { selectRoomInMessageSchema } from "#src/schema/roomsInMessage";
 import { selectUserSchema } from "#src/schema/users";
+import { MODERATION_NOTE_MAX_LENGTH } from "#src/services/message/constants";
 import { createNormalizedStringSchema, getPropertyNames } from "@esposter/shared";
 import { z } from "zod";
-
-export const MODERATION_NOTE_MAX_LENGTH = 2000;
 
 export class ModerationNoteEntity extends AzureEntity {
   declare actorUserId: User["id"];
