+++ packages/db-schema/src/models/resource/ResourceTags.ts
@@ -1,10 +1,7 @@
 import { createNameSchema } from "#src/models/shared/Name";
+import { MAX_TAG_NAME_LENGTH, MAX_TAG_VALUE_LENGTH, MAX_TAGS_COUNT } from "#src/services/resource/constants";
 import { normalizeString } from "@esposter/shared";
 import { z } from "zod";
-
-export const MAX_TAGS_COUNT = 50;
-export const MAX_TAG_NAME_LENGTH = 128;
-export const MAX_TAG_VALUE_LENGTH = 256;
 // Azure tag parity: flat name:value metadata. Names are non-empty; values may be empty
 // ("environment: " is a meaningful Azure tag), so only the name goes through the min(1) name pipe.
 export const resourceTagsSchema = z
