+++ packages/db-schema/src/services/message/webhook/constants.ts
@@ -1,0 +1,11 @@
+export const EMBED_AUTHOR_NAME_MAX_LENGTH = 256;
+// A hex colour code as an integer, so the ceiling is the largest three-byte value.
+export const EMBED_COLOR_MAX_VALUE = 256 ** 3 - 1;
+export const EMBED_DESCRIPTION_MAX_LENGTH = 4096;
+export const EMBED_FIELD_NAME_MAX_LENGTH = 256;
+export const EMBED_FIELD_VALUE_MAX_LENGTH = 1024;
+export const EMBED_FIELDS_MAX_LENGTH = 25;
+export const EMBED_FOOTER_TEXT_MAX_LENGTH = 2048;
+export const EMBED_MAX_LENGTH = 10;
+export const EMBED_TITLE_MAX_LENGTH = 256;
+export const WEBHOOK_CONTENT_MAX_LENGTH = 2000;
